pyproject.toml
[tool.poetry]
name = "aurora-api"
version = "0.1.0"
description = ""
authors = ["akorzunin <akorzunin123@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
hishel = "^0.1.1"
structlog = "^24.4.0"
tortoise-orm = { extras = ["aiosqlite"], version = "^0.23.0" }
aerich = { extras = ["toml"], version = "^0.8.1" }
pydantic = { extras = ["email"], version = "^2.10.5" }


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "C901", # too complex
    "C408", # Unnecessary `dict` call
]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
show_error_codes = true
explicit_package_bases = true

[tool.aerich]
tortoise_orm = "internal.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
